+++ test/api/multipart/multipartUploadPart.spec.ts
@@ -10,7 +10,7 @@ describe('API - multipartUploadPart', () => {
 
   it('should be able to upload part', async() => {
     const settings = getSettingsForTesting({
-      publicKey: factory.publicKey('demo')
+      publicKey: factory.publicKey('image')
     })
     const multipartStartUpload = multipartStart(fileToUpload, settings)
     const {parts} = await multipartStartUpload
@@ -25,7 +25,7 @@ describe('API - multipartUploadPart', () => {
 
   it('should be able to cancel uploading', async(done) => {
     const settings = getSettingsForTesting({
-      publicKey: factory.publicKey('demo')
+      publicKey: factory.publicKey('image')
     })
     const multipartStartUpload = multipartStart(fileToUpload, settings)
     const {parts} = await multipartStartUpload
@@ -45,7 +45,7 @@ describe('API - multipartUploadPart', () => {
 
   it('should be able to handle cancel uploading', async (done) => {
     const settings = getSettingsForTesting({
-      publicKey: factory.publicKey('demo')
+      publicKey: factory.publicKey('image')
     })
     const multipartStartUpload = multipartStart(fileToUpload, settings)
     const {parts} = await multipartStartUpload
@@ -74,7 +74,7 @@ describe('API - multipartUploadPart', () => {
   it('should be able to handle progress', async(done) => {
     let progressValue = 0
     const settings = getSettingsForTesting({
-      publicKey: factory.publicKey('demo')
+      publicKey: factory.publicKey('image')
     })
     const multipartStartUpload = multipartStart(fileToUpload, settings)
     const {parts} = await multipartStartUpload